# Improve performance of `xod-patch-search`