There is no issue.

It reduces search time:
- by ~9% for common searches (like "number", "meter" and etc): 108ms -> 99ms
- by ~54% for search by a library ("lib:xod/core add"): 68ms -> 30ms
- by ~ 91% for search specializations ("when-either-ch(mber"): 74ms -> 6.79ms
- unchanged for search abstracts: 74ms -> 74ms

Also, it exports two new methods: `setCollection` and `getCollection, that will be used in a further optimization.