@brusherru 2018-08-13
@nkrkv @evgenykochetkov check it, please.
It's a small PR, moving search index in the hinting will be in another one.